chore(gateway): add oss to package and module
