+++ refarch-gateway/pom.xml
@@ -9,7 +9,7 @@
         <relativePath />
     </parent>
 
-    <groupId>de.muenchen.refarch</groupId>
+    <groupId>de.muenchen.oss.refarch</groupId>
     <artifactId>refarch-gateway</artifactId>
     <version>1.8.0-SNAPSHOT</version>
