+++ refarch-gateway/src/main/java/de/muenchen/oss/refarch/gateway/configuration/SpaServerCsrfTokenRequestHandler.java
@@ -1,0 +1,41 @@
+package de.muenchen.oss.refarch.gateway.configuration;
+
+import org.springframework.security.web.server.csrf.CsrfToken;
+import org.springframework.security.web.server.csrf.ServerCsrfTokenRequestAttributeHandler;
+import org.springframework.security.web.server.csrf.ServerCsrfTokenRequestHandler;
+import org.springframework.security.web.server.csrf.XorServerCsrfTokenRequestAttributeHandler;
+import org.springframework.web.server.ServerWebExchange;
+import reactor.core.publisher.Mono;
+
+public class SpaServerCsrfTokenRequestHandler extends ServerCsrfTokenRequestAttributeHandler {
+    private final ServerCsrfTokenRequestHandler delegate = new XorServerCsrfTokenRequestAttributeHandler();
+
+    @Override
+    public void handle(final ServerWebExchange exchange, final Mono<CsrfToken> csrfToken) {
+        /*
+         * Always use XorCsrfTokenRequestAttributeHandler to provide BREACH protection of
+         * the CsrfToken when it is rendered in the response body.
+         */
+        this.delegate.handle(exchange, csrfToken);
+    }
+
+    @Override
+    public Mono<String> resolveCsrfTokenValue(final ServerWebExchange exchange, final CsrfToken csrfToken) {
+        /*
+         * If the request contains a request header, use CsrfTokenRequestAttributeHandler
+         * to resolve the CsrfToken. This applies when a single-page application includes
+         * the header value automatically, which was obtained via a cookie containing the
+         * raw CsrfToken.
+         */
+        if (exchange.getRequest().getHeaders().containsKey(csrfToken.getHeaderName())) {
+            return super.resolveCsrfTokenValue(exchange, csrfToken);
+        }
+        /*
+         * In all other cases (e.g. if the request contains a request parameter), use
+         * XorCsrfTokenRequestAttributeHandler to resolve the CsrfToken. This applies
+         * when a server-side rendered form includes the _csrf request parameter as a
+         * hidden input.
+         */
+        return this.delegate.resolveCsrfTokenValue(exchange, csrfToken);
+    }
+}
