+++ refarch-gateway/src/test/java/de/muenchen/oss/refarch/gateway/filter/GlobalAuthenticationErrorFilterTest.java
@@ -1,0 +1,55 @@
+package de.muenchen.oss.refarch.gateway.filter;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
+import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
+import static de.muenchen.oss.refarch.gateway.TestConstants.SPRING_TEST_PROFILE;
+
+import com.github.tomakehurst.wiremock.http.HttpHeader;
+import com.github.tomakehurst.wiremock.http.HttpHeaders;
+import de.muenchen.oss.refarch.gateway.OAuthSecurityMockConfiguration;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
+import org.springframework.context.annotation.Import;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.web.reactive.server.WebTestClient;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ActiveProfiles(SPRING_TEST_PROFILE)
+@AutoConfigureWireMock
+@Import(OAuthSecurityMockConfiguration.class)
+class GlobalAuthenticationErrorFilterTest {
+
+    @Autowired
+    private WebTestClient webTestClient;
+
+    @BeforeEach
+    void setup() {
+        stubFor(get(urlEqualTo("/remote"))
+                .willReturn(aResponse()
+                        .withStatus(HttpStatus.UNAUTHORIZED.value())
+                        .withHeaders(new HttpHeaders(
+                                new HttpHeader("Content-Type", "application/json"),
+                                new HttpHeader("WWW-Authenticate", "Bearer realm=\"Access to the staging site\", charset=\"UTF-8\"")))
+                        .withBody("{ \"testkey\" : \"testvalue\" }")));
+    }
+
+    @Test
+    @WithMockUser
+    void backendAuthenticationError() {
+        webTestClient.get().uri("/api/refarch-gateway-backend-service/remote").exchange()
+                .expectStatus().isEqualTo(HttpStatus.UNAUTHORIZED)
+                .expectHeader().valueMatches("Content-Type", "application/json")
+                .expectHeader().doesNotExist("WWW-Authenticate")
+                .expectBody()
+                .jsonPath("$.status").isEqualTo("401")
+                .jsonPath("$.error").isEqualTo("Authentication Error");
+    }
+
+}
