--- refarch-gateway/src/test/java/de/muenchen/refarch/gateway/filter/GlobalRequestParameterPollutionFilterTest.java
@@ -1,40 +1,0 @@
-package de.muenchen.refarch.gateway.filter;
-
-import static de.muenchen.refarch.gateway.TestConstants.SPRING_TEST_PROFILE;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import de.muenchen.refarch.gateway.OAuthSecurityMockConfiguration;
-import java.nio.charset.StandardCharsets;
-import java.util.Objects;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.web.reactive.server.WebTestClient;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles(SPRING_TEST_PROFILE)
-@Import(OAuthSecurityMockConfiguration.class)
-class GlobalRequestParameterPollutionFilterTest {
-
-    @Autowired
-    private WebTestClient webTestClient;
-
-    @Test
-    @WithMockUser
-    void parameterPollutionAttack() {
-        final StringBuilder jsonResponseBody = new StringBuilder();
-        final String url = "/api/refarch-gateway-backend-service/testendpoint?parameter1=testdata_1&parameter2=testdata&parameter1=testdata_2";
-        webTestClient.get().uri(url).exchange()
-                .expectStatus()
-                .isEqualTo(HttpStatus.BAD_REQUEST)
-                .expectBody()
-                .consumeWith(responseBody -> jsonResponseBody.append(
-                        new String(Objects.requireNonNull(responseBody.getResponseBody()), StandardCharsets.UTF_8)));
-        assertTrue(jsonResponseBody.toString().contains("\"message\":\"parameter pollution\""));
-    }
-
-}
